pyproject.toml
[tool.poetry]
name = "monitoring"
version = "0.1.0"
description = ""
authors = ["Tobias Henkel <tobias.henkel@bmw.de>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pulumi = "3.89.0"
pulumi-docker = "4.4.4"
pulumi-command = "^0.9.1"
pyyaml = "^6.0.1"

[tool.poetry.dev-dependencies]
black = "^23.10.0"
isort = "^5.12.0"
pylint = "^3.0.2"
flake8 = "^6.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
